requirements.txt
streamlit==0.61.0
pandas==0.25.3
numpy==1.18.1
tensorflow-cpu
pathlib
keras==2.4.3
Pillow==8.0.1